Add files via upload
This project employs Support Vector Machine (SVM) algorithm to predict the likelihood of diabetes based on certain health metrics. The dataset is preprocessed and standardized before being split into training and testing sets. The SVM model is trained on the training data and evaluated for accuracy. Finally, the model is used to predict the diabetic status of new input data.